+++ src/main/java/http/Reporter.java
@@ -24,9 +24,9 @@ public class Reporter
         LOGGER.info("Found match client IDs: {}", convertMatchToClientIDsList(match));
     }
 
-    private static List<Integer> convertMatchToClientIDsList(Set<PoolClient> match)
+    private static List<Long> convertMatchToClientIDsList(Set<PoolClient> match)
     {
-        List<Integer> matchedClientsIDs = new ArrayList<>();
+        List<Long> matchedClientsIDs = new ArrayList<>();
         match.forEach(poolClient -> matchedClientsIDs.add(poolClient.getClientID()));
         return matchedClientsIDs;
     }
@@ -52,7 +52,7 @@ public class Reporter
 
     public static void reportMatch(Set<PoolClient> match)
     {
-        List<Integer> clientIDs = convertMatchToClientIDsList(match);
+        List<Long> clientIDs = convertMatchToClientIDsList(match);
 
         RequestBody matchReport = new MultipartBody.Builder()
                 .setType(MultipartBody.FORM)
@@ -71,7 +71,7 @@ public class Reporter
         }
     }
 
-    private static String formatMatchParameter(List<Integer> clientIDs)
+    private static String formatMatchParameter(List<Long> clientIDs)
     {
         StringBuilder encodedListBuilder = new StringBuilder();
         clientIDs.forEach(clientID -> encodedListBuilder.append(clientID).append(';'));
